+++ mongo_perf.py
@@ -294,8 +294,8 @@ def run_program(args_array, func_dict, **kwargs):
     mongo.connect()
 
     # Call function(s) - intersection of command line and function dict.
-    for x in set(args_array.keys()) & set(func_dict.keys()):
-        func_dict[x](mongo, args_array, ofile=outfile, db_tbl=db_tbl,
+    for item in set(args_array.keys()) & set(func_dict.keys()):
+        func_dict[item](mongo, args_array, ofile=outfile, db_tbl=db_tbl,
                      class_cfg=cfg, **kwargs)
 
     cmds_gen.disconnect([mongo])