+++ mongo_perf.py
@@ -224,8 +224,8 @@ def mongo_stat(server, args_array, **kwargs):
             _, value = ast.literal_eval(row).popitem()
             rep_set = value["set"]
             rep_state = value["repl"]
-            value = rm_key(value, "set")
-            value = rm_key(value, "repl")
+            value = gen_libs.rm_key(value, "set")
+            value = gen_libs.rm_key(value, "repl")
             data = {"Server": server.name,
                     "AsOf": gen_libs.get_date() + " " + value["time"],
                     "RepSet": rep_set, "RepState": rep_state,