+++ CHANGELOG.md
@@ -40,6 +40,9 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - mongo_stat:  Reformatted performance stats dictionary, moved "set" and "repl" up one level in dictionary.
 - Documentation updates.
 
+### Removed
+- rm_key function.
+
 
 ## [2.2.0] - 2020-04-08
 ### Added